refactor(Budget) : 당첨금이 클 경우를 대비하기 위해 계산 로직 자료형 long으로 변경

+++ src/main/java/lotto/Domain/Budget.java
@@ -38,7 +38,7 @@ public class Budget {
     }
 
     public double calculateReturnRate(Map<Rank, Integer> rankCount) {
-        int money = 0;
+        long money = 0;
         for (Rank rank : Rank.values()) {
             money += rank.getPrizeAmount() * rankCount.getOrDefault(rank, 0);
         }
